Shop buying works now
Submitting what I have done as I likely won't be able to do anything for a few days. Ideas on what to do next, selling (remove quota system and just allow them to buy whatever is on the boat), adding an item system that adds all the possible items at the start of the game, event implementation , and end conditions.