Created event logic
Made some changes to items, need to create logic for special items, Events has fight logic, game will need to update to repair the ship, the events need to be weighted more, rewards have to be added. Changed the logic for ending the game to provide better parity with ship detruction.