Quality changes and fixes to meet specifications
Added option to surrender items to pirates and flee after engaging in a fight. Moved event to be a component of route. Added different routes having different event chances. Changed an element in game to accept a route as a variable. Added confirm to exit. Added a variable to keep track of the purchase price of items and increased the sell price of all items. Moved dump from event to player. Added more tests for Card, Event, Game, and Player